helpers: split assert and getter events files
