+++ test/contracts/apps/recovery_to_vault.js
@@ -2,7 +2,8 @@ const { hash } = require('eth-ens-namehash')
 const { getBalance } = require('../../helpers/web3')
 const { skipCoverage } = require('../../helpers/coverage')
 const { assertRevert } = require('../../helpers/assertThrow')
-const { assertAmountOfEvents, assertEvent, getNewProxyAddress } = require('../../helpers/assertEvent')(web3)
+const { getNewProxyAddress } = require('../../helpers/events')
+const { assertAmountOfEvents, assertEvent } = require('../../helpers/assertEvent')(web3)
 
 const ACL = artifacts.require('ACL')
 const Kernel = artifacts.require('Kernel')
