+++ test/contracts/evmscript/evm_script.js
@@ -3,7 +3,8 @@ const { rawEncode } = require('ethereumjs-abi')
 const { soliditySha3 } = require('web3-utils')
 const { assertRevert } = require('../../helpers/assertThrow')
 const { createExecutorId, encodeCallScript } = require('../../helpers/evmScript')
-const { assertEvent, assertAmountOfEvents, getEventArgument, getNewProxyAddress } = require('../../helpers/assertEvent')(web3)
+const { assertEvent, assertAmountOfEvents } = require('../../helpers/assertEvent')(web3)
+const { getEventArgument, getNewProxyAddress } = require('../../helpers/events')
 
 const ACL = artifacts.require('ACL')
 const Kernel = artifacts.require('Kernel')
