+++ test/contracts/kernel/kernel_acl.js
@@ -2,7 +2,8 @@ const { hash } = require('eth-ens-namehash')
 const { keccak_256 } = require('js-sha3')
 const { soliditySha3 } = require('web3-utils')
 const { assertRevert } = require('../../helpers/assertThrow')
-const { assertAmountOfEvents, getEventArgument, getNewProxyAddress, assertEvent } = require('../../helpers/assertEvent')(web3)
+const { assertAmountOfEvents, assertEvent } = require('../../helpers/assertEvent')(web3)
+const { getEventArgument, getNewProxyAddress } = require('../../helpers/events')
 
 const ACL = artifacts.require('ACL')
 const Kernel = artifacts.require('Kernel')
